+++ packages/jest-config/src/importMjs.ts
@@ -5,6 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import {pathToFileURL} from 'url';
+
 // this is in a separate file so that node 8 don't explode with a syntax error.
 // Remove this file when we drop support for Node 8
-export default (specifier: string) => import(specifier);
+export default (specifier: string): Promise<{default: unknown}> =>
+  // node `import()` supports URL, but TypeScript doesn't know that
+  import(pathToFileURL(specifier).href);