+++ packages/jest-core/src/cli/index.ts
@@ -31,13 +31,13 @@ const {print: preRunMessagePrint} = preRunMessage;
 
 type OnCompleteCallback = (results: AggregatedResult) => void;
 
-export const runCLI = async (
+export async function runCLI(
   argv: Config.Argv,
   projects: Array<Config.Path>,
 ): Promise<{
   results: AggregatedResult;
   globalConfig: Config.GlobalConfig;
-}> => {
+}> {
   const realFs = require('fs');
   const fs = require('graceful-fs');
   fs.gracefulify(realFs);
@@ -109,7 +109,7 @@ export const runCLI = async (
   }
 
   return {globalConfig, results};
-};
+}
 
 const buildContextsAndHasteMaps = async (
   configs: Array<Config.ProjectConfig>,