+++ packages/jest-diff/src/printDiffs.ts
@@ -102,7 +102,10 @@ export const printCommonLine = (
     emptyFirstOrLastLinePlaceholder,
   );
 
-export const hasCommonDiff = (diffs: Array<Diff>, isMultiline: boolean) => {
+export const hasCommonDiff = (
+  diffs: Array<Diff>,
+  isMultiline: boolean,
+): boolean => {
   if (isMultiline) {
     // Important: Ignore common newline that was appended to multiline strings!
     const iLast = diffs.length - 1;