+++ packages/jest-reporters/src/coverage_reporter.ts
@@ -57,7 +57,7 @@ export default class CoverageReporter extends BaseReporter {
     this._options = options || {};
   }
 
-  onTestResult(_test: Test, testResult: TestResult) {
+  onTestResult(_test: Test, testResult: TestResult): void {
     if (testResult.v8Coverage) {
       this._v8CoverageResults.push(testResult.v8Coverage);
       return;
@@ -93,7 +93,7 @@ export default class CoverageReporter extends BaseReporter {
   async onRunComplete(
     contexts: Set<Context>,
     aggregatedResults: AggregatedResult,
-  ) {
+  ): Promise<void> {
     await this._addUntestedFiles(contexts);
     const {map, reportContext} = await this._getCoverageResult();