+++ packages/jest-validate/src/utils.ts
@@ -14,12 +14,12 @@ export const DEPRECATION = `${BULLET} Deprecation Warning`;
 export const ERROR = `${BULLET} Validation Error`;
 export const WARNING = `${BULLET} Validation Warning`;
 
-export const format = (value: any): string =>
+export const format = (value: unknown): string =>
   typeof value === 'function'
     ? value.toString()
     : prettyFormat(value, {min: true});
 
-export const formatPrettyObject = (value: any): string =>
+export const formatPrettyObject = (value: unknown): string =>
   typeof value === 'function'
     ? value.toString()
     : JSON.stringify(value, null, 2)
@@ -43,7 +43,7 @@ export const logValidationWarning = (
   name: string,
   message: string,
   comment?: string | null,
-) => {
+): void => {
   comment = comment ? '\n\n' + comment : '\n';
   console.warn(chalk.yellow(chalk.bold(name) + ':\n\n' + message + comment));
 };
@@ -51,7 +51,7 @@ export const logValidationWarning = (
 export const createDidYouMeanMessage = (
   unrecognized: string,
   allowedOptions: Array<string>,
-) => {
+): string => {
   const suggestion = allowedOptions.find(option => {
     const steps: number = leven(option, unrecognized);
     return steps < 3;